Backend/requirement.txt
asgiref
Django
django-cors-headers
djangorestframework
django-ratelimit
djangorestframework-simplejwt
PyJWT
Pillow
pytz
sqlparse
psycopg2-binary
python-dotenv
django-quill-editor